requirements.txt
numpy==1.19.2
SQLAlchemy==1.3.20
pydantic==1.8.2
pandas==1.1.3
requests==2.22.0
pytrends==4.7.3
fastapi==0.65.1
PyYAML==5.4.1
uvicorn==0.13.4